Move `krel changelog` logic into separate package
This allows us to re-use the package later on in `krel stage`.
